refactor: Refactor to separate view and logic of create card component
